Move task callback to avoid circular imports
