+++ handler/handler.go
@@ -1,7 +1,6 @@
 package handler
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"html/template"
@@ -80,51 +79,6 @@ func GetTaskInfo(scheduler types.Scheduler) http.HandlerFunc {
 			writeJSON(http.StatusOK, task, w)
 		}
 	}
-}
-
-type callbackData struct {
-	Time   int64  `json:"time"`
-	Status string `json:"status"`
-	TaskID string `json:"task_id"`
-}
-
-// NotifyCallback handles posting a JSON back to the URI given with the task.
-func NotifyCallback(task *types.EremeticTask) {
-	if len(task.CallbackURI) == 0 {
-		return
-	}
-
-	cbData := callbackData{
-		Time:   task.Status[len(task.Status)-1].Time,
-		Status: task.Status[len(task.Status)-1].Status,
-		TaskID: task.ID,
-	}
-
-	body, err := json.Marshal(cbData)
-	if err != nil {
-		logrus.WithError(err).WithFields(logrus.Fields{
-			"task_id":      task.ID,
-			"callback_uri": task.CallbackURI,
-		}).Error("Unable to create callback message")
-		return
-	}
-
-	go func() {
-		_, err = http.Post(task.CallbackURI, "application/json", bytes.NewBuffer(body))
-
-		if err != nil {
-			logrus.WithError(err).WithFields(logrus.Fields{
-				"task_id":      task.ID,
-				"callback_uri": task.CallbackURI,
-			}).Error("Unable to POST to Callback URI")
-		} else {
-			logrus.WithFields(logrus.Fields{
-				"task_id":      task.ID,
-				"callback_uri": task.CallbackURI,
-			}).Debug("Sent callback")
-		}
-	}()
-
 }
 
 func handleError(err error, w http.ResponseWriter, message string) {
